events: move the daemon-side machinery into daemon::events
The EventBus, sink, listener, socket-path convention, and their tests
are daemon internals, not part of the `events` subcommand. Move them to
libshpool/src/daemon/events.rs; libshpool/src/events.rs is now just the
subcommand (run). The daemon owns socket_path (it binds the socket) and
the subcommand follows it, so the dependency points down: lib.rs wires
events::run to crate::daemon::events::socket_path. Repoint the
daemon::events references in daemon/mod.rs, server.rs, and ttl_reaper.rs;
behavior is unchanged.
